backend/requirements.txt
autogen-agentchat
autogen-core
autogen-ext[mcp]
autogen-ext[chromadb]
fastapi
uvicorn[standard]
requests
python-dotenv
tavily-python
openai
websockets
chromadb
grpcio>=1.67.0
grpcio-tools>=1.67.0
